Get connection to DB and read data
This commit reads data from the CSV file.
In order to get a connection to the DB, you
will need to get the connection jdbc for postgresql
and it to your reference library through your
prefered IDE. You will also need to create your own
password and username to connect to the Database.